pyproject.toml
[project]
name = "class-7-assignment"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chainlit>=2.7.2",
    "openai-agents>=0.2.10",
]
